feat: send emotes finally to the foreground app.
This is still a bit broken and buggy. There's a panic condition on when
the emote candidate list is too large, or when `emote_index` goes
beyonds the bounds of the vector.

There's also a race condition on when you open a window. Because thats
not an instand operation, theres a fair bit of time where the window
is in the process of opening (`self.win` isn't set yet), along with the
process listening for another window spawn toggle.